Add support for callback factories
This commit refactors the Builder API to accept a MakeCallback impl
instead of a callback. This trait can be invoked by the crate to convery
an arbitrary type into a status line callback.

By putting a trait bound on the external API the crate can now be used
in two ways:

1. A factory can be provided to more easily couple (shared) state to the
   status line callback.

2. By leveraging a generic impl on `FnMut(&mut W) -> Result<u16>` all
   compatible closures automatically implement the factory trait to emit
   Self, allowing a closure or free function to be passed in without
   creating a factory. This allows for a simplified interface where the
   callback is provided directly.

Additionally, the signature of the callback was simplified to provide a
`&mut W` with `W: Write` instead of requiring that `&W: Write`, which is
not as flexible.